# ddtrace/tracer: check if value implements fmt.Stringer in span.SetTag