It avoids to call fmt.Sprint, which uses reflection and is slower.

Fixes https://github.com/DataDog/dd-trace-go/issues/798

Benchmark:

```
go test -v -run=^$ -bench=^BenchmarkSetTag -benchmem -benchtime=10s -cpu=1 ./ddtrace/tracer

benchstat -delta-test none old.txt new.txt
name            old time/op    new time/op    delta
SetTagMetric      63.3ns ± 0%    64.8ns ± 0%   +2.37%
SetTagString      61.8ns ± 0%    61.4ns ± 0%   -0.65%
SetTagStringer     247ns ± 0%      77ns ± 0%  -68.87%
SetTagField       26.8ns ± 0%    26.8ns ± 0%    0.00%

name            old alloc/op   new alloc/op   delta
SetTagMetric       4.00B ± 0%     4.00B ± 0%    0.00%
SetTagString       4.00B ± 0%     4.00B ± 0%    0.00%
SetTagStringer     16.0B ± 0%      4.0B ± 0%  -75.00%
SetTagField        0.00B          0.00B         0.00%

name            old allocs/op  new allocs/op  delta
SetTagMetric        1.00 ± 0%      1.00 ± 0%    0.00%
SetTagString        1.00 ± 0%      1.00 ± 0%    0.00%
SetTagStringer      2.00 ± 0%      1.00 ± 0%  -50.00%
SetTagField         0.00           0.00         0.00%
```